Add more error checking into getGeosFp and doGeos57Multi
perl/getGeosFp:
(1) Compute the # of files successfully downloaded
    at the end of the script.  Print this to the stdout.
(2) Also do not exit if the listing file is not found, but
    re-download it instead.

perl/doGeos57Multi: Put in a basic error check that will prevent
the job from starting if there are too few files present.  This is not
an exhaustive check but it should prevent the met field code from
executing prematurely when submitting jobs via the at command.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>