Add better checks for missing values into adjust_all.ncl
In the adjust/adjust_all.ncl file, if the file list command

  files = systemfunc( "ls -1 " + dir + "GEOS572*A1*.nc" )

(and similar routines) cannot find any files in the given
directory, then the files array will have a missing value.
We need to use the function ismissing() to screen out these
missing values.

Furthermore, we now also use the reassignment operator:

  files := systemfunc( "ls -1 " + dir + "GEOS572*A1*.nc" )

so that it will be resized properly if the number of files
found changes.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>