Modify moveGeosFp to skip the temporary directory
We have modified moveGeosFp so that it will skip moving
processed files (CH, EU, NA, SE, 2x25, 4x5) if the temporary
directory is the same as the data directory.  This allows you
to write directly into the data directory.

NOTE: In bin/GeosFpDriver.input, if you wish to omit using
the temporary directory, you must use the same string for
both temporary directory and data directory, for example:

  ==> 4 x 5 output
  T
  GEOSFP.YYYYMMDD.%%%%%%.4x5.nc
  /mnt/gcgrid-rw/GEOS_4x5/GEOS_FP/YYYY/MM/
  /mnt/gcgrid-rw/GEOS_4x5/GEOS_FP/YYYY/MM/

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>