Now restore CLOUD from the tavg3_3d_rad_Nv collection
In the GEOS_FP code we had computed CLOUD with the formula:

  CLOUD = min( CFAN + CFLS, 1.0 )

However, this approximation leads to cloud fractions that
are higher, which leads to more O3 and OH in GEOS-Chem
simulations.

Here we have now restored obtaining and regridding CLOUD from
the tavg3_3d_rad_Nv collection.  We also have created a new
script and code to just reprocess the A3cld file without
having to reprocess the entire met archive.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>